+++ lang/fr/modal.php
@@ -1,6 +1,7 @@
 <?php
 
 return [
+    'confirm' => 'Veuillez confirmer',
     'close' => 'Fermer',
     'yes' => 'Oui',
     'no' => 'Non',