+++ resources/views/bootstrap-5/components/forms/inputs/password.blade.php
@@ -1,10 +1,5 @@
-<input
-    name="{{ $name }}"
-    type="password"
-    id="{{ $id }}"
-    @if($hasErrors)aria-describedby="validation-{{ $name }}-feedback"@endif
-    {{ $attributes->class([
-        'form-control',
-        'is-invalid' => $hasErrors,
-    ]) }}
-/>
+{{--
+    This view serves no more purpose than that of its parent component.
+    It is only there so that it can be published and overloaded.
+--}}
+@include('blade-ui-kit-bootstrap::bootstrap-5.components.forms.inputs.input')