+++ src/Components/Buttons/FormButton.php
@@ -5,19 +5,21 @@ declare(strict_types=1);
 namespace BladeUIKitBootstrap\Components\Buttons;
 
 use BladeUIKitBootstrap\Components\BladeComponent;
-use BladeUIKitBootstrap\Concerns\HasFormMethod;
+use BladeUIKitBootstrap\Concerns\FormMethod;
 use Illuminate\Support\Str;
 
 class FormButton extends BladeComponent
 {
-    use HasFormMethod;
+    use FormMethod;
 
     public function __construct(
-        public ?string $formId = null,
         public ?string $action = null,
+        public ?string $formId = null,
+        public ?string $title = null,
+        public ?string $confirm = null,
+        public bool $disabled = false,
+        public bool $novalidate = true,
         string $method = 'POST',
-        public bool $hasFiles = false,
-        public bool $novalidate = true
     ) {
         $this->formId = 'form-button-'.($formId ?? Str::random(32));
         $this->method = $this->validFormMethod($method);