+++ src/Components/Forms/Inputs/Text.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BladeUIKitBootstrap\Components\Forms\Inputs;
+
+class Text extends Input
+{
+    public function __construct(
+        string $name,
+        string $value = null,
+        string $id = null,
+        string $errorBag = null,
+    ) {
+        parent::__construct(
+            name: $name,
+            type: 'text',
+            value: $value,
+            id: $id,
+            errorBag: $errorBag,
+        );
+    }
+
+    public function viewName(): string
+    {
+        return 'components.forms.inputs.email';
+    }
+}