+++ src/Components/Forms/Label.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace BladeUIKitBootstrap\Components\Forms;
 
 use BladeUIKitBootstrap\Components\BladeComponent;
-use Illuminate\Support\Str;
 
 class Label extends BladeComponent
 {
@@ -16,7 +15,7 @@ class Label extends BladeComponent
 
     public function fallback(): string
     {
-        return Str::ucfirst(str_replace('_', ' ', $this->for));
+        return \str($this->for)->replace('_', ' ')->ucfirst->toString();
     }
 
     public function viewName(): string