+++ src/Concerns/FormMethod.php
@@ -6,7 +6,7 @@ namespace BladeUIKitBootstrap\Concerns;
 
 use InvalidArgumentException;
 
-trait HasFormMethod
+trait FormMethod
 {
     public string $method = 'POST';
 
@@ -31,7 +31,7 @@ trait HasFormMethod
             throw new InvalidArgumentException(\sprintf(
                 'The HTTP method "%s" is not allowed. Allowed method are: %s.',
                 e($method),
-                implode(', ', self::ALLOWED_FORM_METHOD)
+                \implode(', ', self::ALLOWED_FORM_METHOD)
             ));
         }