+++ src/Http/Middleware/BladeUiKitBootstrap4.php
@@ -11,11 +11,11 @@ class BladeUiKitBootstrap4
     /**
      * Handle an incoming request.
      *
-     * @param  \Closure(\Illuminate\Http\Request): (\Symfony\Component\HttpFoundation\Response)  $next
+     * @param  \Closure(Request):Response  $next
      */
     public function handle(Request $request, Closure $next): Response
     {
-        app('config')->set(['blade-ui-kit-bootstrap.boostrap_version' => 'bootstrap-4']);
+        \app('config')->set(['blade-ui-kit-bootstrap.boostrap_version' => 'bootstrap-4']);
 
         return $next($request);
     }